# lib/units/api/swagger/api_v1.yaml
swagger: "2.0"
info:
  version: "2.0.0"
  title: Smartphone Test Farm
  description: Control and manages real Smartphone devices from browser and restful apis
  license:
    name: Apache-2.0
    url: http://www.apache.org/licenses/LICENSE-2.0
  contact:
    name: STF Team
    email: contact@openstf.io
    url: http://openstf.io/
basePath: /api/v1
schemes:
  - http
  - https
consumes:
  - application/json
produces:
  - application/json
...
paths:
...
  /schedules/{serial}:
    x-swagger-router-controller: schedule
    get:
      summary: Schedule Information
      description: The schedule enpoint return information list of specified device schedules
      operationId: getSchedulesBySerial
      tags:
        - devices
      parameters:
        - name: serial
          in: path
          description: Device Serial
          required: true
          type: string
        - name: start
          in: query
          description: Fields query parameter takes utc timestamp. The schedules will be filtered by the end time later than specified timestamp.
          required: false
          type: integer
        - name: end
          in: query
          description: Fields query parameter takes utc timestamp. The schedules will be filtered by the end time earlier than specified timestamp.
          required: false
          type: integer
      responses:
        "200":
          description: List of Schedules
          schema:
            $ref: "#/definitions/ScheduleListResponse"
        default:
          description: Unexpected Error
          schema:
            $ref: "#/definitions/ErrorResponse"
      security:
        - accessTokenAuth: []
    post:
      summary: Add new schedule Information
      description: The schedule enpoint register new reservation schedule
      operationId: addSchedule
      tags:
        - schedules
      parameters:
        - name: serial
          in: path
          description: Device Serial
          required: true
          type: string
        - name: schedule
          in: body
          description: Fields schedule parameter takes reservation data.
          required: true
          schema:
            $ref: "#/definitions/ScheduleRequest"
      responses:
        "201":
          description: Registerd schedule
          schema:
            $ref: "#/definitions/ScheduleResponse"
        default:
          description: Unexpected Error
          schema:
            $ref: "#/definitions/ErrorResponse"
      security:
        - accessTokenAuth: []
  /schedules/{serial}/{id}:
    x-swagger-router-controller: schedule
    put:
      summary: Update schedule
      description: The schedule enpoint updates specified reservation schedule
      operationId: updateSchedule
      tags:
        - schedules
      parameters:
        - name: serial
          in: path
          description: Device Serial
          required: true
          type: string
        - name: id
          in: path
          description: Schedule id
          required: true
          type: string
        - name: schedule
          in: body
          description: Fields schedule parameter takes reservation data.
          required: true
          schema:
            $ref: "#/definitions/ScheduleRequest"
      responses:
        "201":
          description: Registerd schedule
          schema:
            $ref: "#/definitions/ScheduleResponse"
        default:
          description: Unexpected Error
          schema:
            $ref: "#/definitions/ErrorResponse"
      security:
        - accessTokenAuth: []
    delete:
      summary: Delete schedule
      description: The schedule enpoint removes specified reservation schedule
      operationId: deleteSchedule
      tags:
        - schedules
      parameters:
        - name: serial
          in: path
          description: Device Serial
          required: true
          type: string
        - name: id
          in: path
          description: Schedule id
          required: true
          type: string
      responses:
        "204":
          description: No Contents
        default:
          description: Unexpected Error
          schema:
            $ref: "#/definitions/ErrorResponse"
      security:
        - accessTokenAuth: []
...
definitions:
...
  ScheduleListResponse:
    required:
      - schedules
    properties:
      schedules:
        type: array
        items:
          type: object
  ScheduleRequest:
    required:
      - start
      - end
    properties:
      start:
        description: The start time utc timestamp in ms.
        type: integer
      end:
        description: The end time utc timestamp in ms.
        type: integer
  ScheduleResponse:
    required:
      - schedule
    properties:
      schedule:
        type: object
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string
securityDefinitions:
  accessTokenAuth:
    type: apiKey
    name: authorization
    in: header